+++ apps/operating-graph/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atrib/operating-graph
 
+## 0.1.14
+
+### Patch Changes
+
+- @atrib/sdk@7.1.2
+
 ## 0.1.13
 
 ### Patch Changes